requirements.txt
aiohttp>=3.9.0
Pillow>=10.0.0
PyYAML>=6.0.0
